64 bit compatibility effort, 1/nn: type correctness work in common.c, common.h, gl_vidsdl.c, mathlib.h, zone.c, zone.h. added Z_Realloc as a new function (will be used later.) next step will be type correctness work in opengl stuff.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@33 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/common.c
@@ -150,11 +150,11 @@ void InsertLinkAfter (link_t *l, link_t *after)
 ============================================================================
 */
 
-void Q_memset (void *dest, int fill, int count)
+void Q_memset (void *dest, int fill, size_t count)
 {
-	int             i;
+	size_t		i;
 
-	if ( (((long)dest | count) & 3) == 0)
+	if ( (((size_t)dest | count) & 3) == 0)
 	{
 		count >>= 2;
 		fill = fill | (fill<<8) | (fill<<16) | (fill<<24);
@@ -166,11 +166,11 @@ void Q_memset (void *dest, int fill, int count)
 			((byte *)dest)[i] = fill;
 }
 
-void Q_memcpy (void *dest, const void *src, int count)
+void Q_memcpy (void *dest, const void *src, size_t count)
 {
-	int             i;
+	size_t		i;
 
-	if (( ( (long)dest | (long)src | count) & 3) == 0 )
+	if (( ( (size_t)dest | (size_t)src | count) & 3) == 0 )
 	{
 		count>>=2;
 		for (i=0 ; i<count ; i++)
@@ -181,7 +181,7 @@ void Q_memcpy (void *dest, const void *src, int count)
 			((byte *)dest)[i] = ((byte *)src)[i];
 }
 
-int Q_memcmp (const void *m1, const void *m2, int count)
+int Q_memcmp (const void *m1, const void *m2, size_t count)
 {
 	while(count)
 	{
