+++ Quake/common.h
@@ -124,9 +124,9 @@ float MSG_ReadAngle16 (void); //johnfitz
 
 //============================================================================
 
-void Q_memset (void *dest, int fill, int count);
-void Q_memcpy (void *dest, const void *src, int count);
-int Q_memcmp (const void *m1, const void *m2, int count);
+void Q_memset (void *dest, int fill, size_t count);
+void Q_memcpy (void *dest, const void *src, size_t count);
+int Q_memcmp (const void *m1, const void *m2, size_t count);
 void Q_strcpy (char *dest, const char *src);
 void Q_strncpy (char *dest, const char *src, int count);
 int Q_strlen (const char *str);
